requirements.txt
pandas==1.3.5
requests==2.26.0
python-dotenv==0.21.0
google-api-python-client==2.50.0
supabase==0.2.6